+++ kernel/agent_policy.c
@@ -79,9 +79,11 @@ void agent_policy_load_defaults(void) {
     memcpy(allow_tools[9], "fs.rm", 6);
     memcpy(allow_tools[10], "fs.search", 10);
     memcpy(allow_tools[11], "fs.grep", 8);
-    memcpy(allow_tools[12], "sys.info", 9);
-    memcpy(allow_tools[13], "net.ping", 9);
-    allow_tool_count = 14;
+    memcpy(allow_tools[12], "fs.diff", 8);
+    memcpy(allow_tools[13], "sys.info", 9);
+    memcpy(allow_tools[14], "net.ping", 9);
+    memcpy(allow_tools[15], "net.fetch", 10);
+    allow_tool_count = 16;
 }
 
 void agent_policy_reload(void) {